testing: move BenchmarkContracts
