+++ persist/postgres/contracts_test.go
@@ -3,6 +3,7 @@ package postgres
 import (
 	"context"
 	"errors"
+	"fmt"
 	"reflect"
 	"testing"
 	"time"
@@ -11,6 +12,7 @@ import (
 	"go.sia.tech/coreutils/chain"
 	"go.sia.tech/indexd/contracts"
 	"go.sia.tech/indexd/subscriber"
+	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
 	"lukechampine.com/frand"
 )
@@ -780,4 +782,103 @@ func TestSyncContract(t *testing.T) {
 		Size:               1900,
 		UsedCollateral:     types.Siacoins(20),
 	})
+}
+
+// BenchmarkContracts is a benchmark to ensure the performance of
+// Contract and Contracts.
+//
+// M1 Max | Contract       |                | 1 ms/op
+// M1 Max | Contracts 100  | None           | 1.9 ms/op
+// M1 Max | Contracts 100  | Revisable      | 2.4 ms/op
+// M1 Max | Contracts 100  | Revisable+Good | 2.2 ms/op
+// M1 Max | Contracts 1000 | None           | 5.1 ms/op
+// M1 Max | Contracts 1000 | Revisable      | 4.8 ms/op
+// M1 Max | Contracts 1000 | Revisable+Good | 5.3 ms/op
+func BenchmarkContracts(b *testing.B) {
+	const (
+		numContractsPerHost = 100
+		numHosts            = 500
+	)
+
+	// prepare database
+	store := initPostgres(b, zap.NewNop())
+	hosts := make([]types.PublicKey, 0, numHosts)
+	contractIDs := make([]types.FileContractID, 0, numHosts*numContractsPerHost)
+	if err := store.transaction(context.Background(), func(ctx context.Context, tx *txn) error {
+		for range numHosts {
+			var hostID int64
+			hk := types.GeneratePrivateKey().PublicKey()
+			err := tx.QueryRow(ctx, `INSERT INTO hosts (public_key, last_announcement) VALUES ($1, NOW()) RETURNING id;`, sqlPublicKey(hk)).Scan(&hostID)
+			if err != nil {
+				return err
+			}
+
+			hostContractIDs := make([]types.FileContractID, numContractsPerHost)
+			for i := range numContractsPerHost {
+				frand.Read(hostContractIDs[i][:])
+				if _, err := tx.Exec(ctx, `INSERT INTO contracts (host_id, contract_id, proof_height, expiration_height, contract_price, initial_allowance, miner_fee, total_collateral, remaining_allowance, state, good) VALUES ($1, $2, 0, 0, $3, $4, $5, $6, $7, $8, $9);`,
+					hostID,
+					sqlHash256(hostContractIDs[i][:]),
+					sqlCurrency(types.ZeroCurrency),
+					sqlCurrency(types.ZeroCurrency),
+					sqlCurrency(types.ZeroCurrency),
+					sqlCurrency(types.ZeroCurrency),
+					sqlCurrency(types.ZeroCurrency),
+					sqlContractState(uint8(frand.Uint64n(5))), // 40% active
+					frand.Uint64n(2) == 0,                     // 50% good
+				); err != nil {
+					return err
+				}
+			}
+			for i := 1; i < len(hostContractIDs)-1; i++ {
+				if _, err := tx.Exec(ctx, `UPDATE contracts SET renewed_from = $1, renewed_to = $2 WHERE contract_id = $3;`, hostContractIDs[i-1], hostContractIDs[i+1], hostContractIDs[i]); err != nil {
+					return err
+				}
+			}
+			contractIDs = append(contractIDs, hostContractIDs...)
+			hosts = append(hosts, hk)
+		}
+		return nil
+	}); err != nil {
+		b.Fatal(err)
+	}
+
+	b.Run("contract", func(b *testing.B) {
+		if b.N > len(contractIDs) {
+			b.Fatalf("too many iterations, %d > %d", b.N, len(contractIDs))
+		}
+		for b.Loop() {
+			_, err := store.Contract(context.Background(), contractIDs[b.N%len(contractIDs)])
+			if err != nil {
+				b.Fatal(err)
+			}
+		}
+	})
+
+	for _, limit := range []int{100, 1000} {
+		b.Run(fmt.Sprintf("contracts_%d", limit), func(b *testing.B) {
+			for b.Loop() {
+				_, err := store.Contracts(context.Background(), 0, limit)
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+		b.Run(fmt.Sprintf("contracts_revisable_%d", limit), func(b *testing.B) {
+			for b.Loop() {
+				_, err := store.Contracts(context.Background(), 0, limit, contracts.WithRevisable(true))
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+		b.Run(fmt.Sprintf("contracts_revisable+good_%d", limit), func(b *testing.B) {
+			for b.Loop() {
+				_, err := store.Contracts(context.Background(), 0, limit, contracts.WithRevisable(true), contracts.WithGood(true))
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+	}
 }
